+++ ClientLib/src/main/java/org/droidplanner/services/android/impl/core/drone/autopilot/generic/GenericMavLinkDrone.java
@@ -237,12 +237,13 @@ public class GenericMavLinkDrone implements MavLinkDrone {
     }
 
     @Override
-    public byte getSysid() {
+    public short getSysid() {
+
         return heartbeat.getSysid();
     }
 
     @Override
-    public byte getCompid() {
+    public short getCompid() {
         return heartbeat.getCompid();
     }
